add: mypage info edit profile component
