+++ src/components/Modal/Profile/MyTeamProfileModal.js
@@ -1,11 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import { Modal } from 'antd';
 import { ReactComponent as UniversityMark } from '../../../asset/svg/UniversityMark.svg';
 import { ReactComponent as UniversityMarkGray } from '../../../asset/svg/UniversityMarkGray.svg';
-import { ReactComponent as Info } from '../../../asset/svg/Info.svg';
-import ApplyButton from '../../Button/ApplyButton';
-import ModifyProfileModal from './ModifyProfileModal';
 import SliderBoxMembers from '../../Slider/SliderBoxMembers';
 import AreaText from '../../MainRecommend/AreaText';
 import DateText from '../../MainRecommend/DateText';
@@ -15,8 +12,7 @@ import {
 } from '../../../features/api/userApi';
 import DrinkText from '../../MainRecommend/DrinkText';
 import MemberText from '../../MainRecommend/MemberText';
-import DeleteMyProfileModal from './DeleteMyProfileModal';
-import NoticeModal from './NoticeModal';
+import EditProfile from '../../MainRecommend/EditProfile';
 
 export default function MyTeamProfileModal(props) {
   const { open, setModal } = props;
@@ -24,10 +20,6 @@ export default function MyTeamProfileModal(props) {
   const { data: myProfile } = useGetProfileQuery(myTeamId, {
     skip: !myTeamId,
   });
-
-  const [isModifyModalOpen, setIsModifyModalOpen] = useState(false);
-  const [isStopMatchingModalOpen, setIsStopMatchingModalOpen] = useState(false);
-  const [isNoticeModalOpen, setIsNoticeModalOpen] = useState(false);
 
   useEffect(() => {
     setModal(false);
@@ -45,92 +37,55 @@ export default function MyTeamProfileModal(props) {
           closable
           onCancel={() => setModal(false)}
         >
-          <ModifyProfileModal
-            open={isModifyModalOpen}
-            setModal={() => {
-              setIsModifyModalOpen((prev) => !prev);
-            }}
-          />
-          <DeleteMyProfileModal
-            open={isStopMatchingModalOpen}
-            setModal={() => {
-              setIsStopMatchingModalOpen((prev) => !prev);
-            }}
-            teamId={myTeamId}
-          />
-          <NoticeModal
-            open={isNoticeModalOpen}
-            setModal={() => {
-              setIsNoticeModalOpen((prev) => !prev);
-            }}
-          />
           {myProfile && (
-            <>
-              <TeamProfile>
-                <TeamIntro>
-                  <TeamName>{myProfile.teamName}</TeamName>
-                  <EditBtn onClick={() => setIsModifyModalOpen(true)}>
-                    수정
-                  </EditBtn>
-                  <EditBtn onClick={() => setIsStopMatchingModalOpen(true)}>
-                    삭제
-                  </EditBtn>
-                  <SInfo onClick={() => setIsNoticeModalOpen(true)} />
-                </TeamIntro>
-                <TextBox>
-                  <Title>한 줄 어필</Title>
-                  <Content>{myProfile.intro}</Content>
-                </TextBox>
-                <TextBox>
-                  <Container>
-                    <Title>기본 정보</Title>
-                    {myProfile.approval ? (
-                      <>
-                        <SUniversityMark />
-                        <UniversityMarkText>대학 인증 완료</UniversityMarkText>
-                      </>
-                    ) : (
-                      <>
-                        <SUniversityMarkGray />
-                        <UniversityNoMarkText>대학 미인증</UniversityNoMarkText>
-                      </>
-                    )}
-                  </Container>
-                  <TeamInfo>
-                    <Subtitle>일정</Subtitle>
-                    <SubContent>
-                      <DateText availableDates={myProfile.teamAvailableDate} />
-                    </SubContent>
-                  </TeamInfo>
-                  <TeamInfo>
-                    <Subtitle>지역</Subtitle>
-                    <AreaText areaProps={myProfile.areas} />
-                  </TeamInfo>
-                  <TeamInfo>
-                    <Subtitle>주량</Subtitle>
-                    <DrinkText drink={myProfile.drink} />
-                  </TeamInfo>
-                  <TeamInfo>
-                    <Subtitle>인원</Subtitle>
-                    <MemberText
-                      count={myProfile.memberCount}
-                      more={myProfile.memberCounts}
-                    />
-                  </TeamInfo>
-                </TextBox>
-                <SliderBoxMembers members={myProfile.members} />
-              </TeamProfile>
-              {/* <Footer> */}
-              {/* <ButtonBox> */}
-              {/* <ApplyButton onClick={() => setIsStopMatchingModalOpen(true)}>
-                    매칭 중단하기
-                  </ApplyButton> */}
-              {/* <ApplyButton2 onClick={() => setIsModifyModalOpen(true)}>
-                  수정하기
-                </ApplyButton2> */}
-              {/* </ButtonBox> */}
-              {/* </Footer> */}
-            </>
+            <TeamProfile>
+              <TeamIntro>
+                <TeamName>{myProfile.teamName}</TeamName>
+                <EditProfile />
+              </TeamIntro>
+              <TextBox>
+                <Title>한 줄 어필</Title>
+                <Content>{myProfile.intro}</Content>
+              </TextBox>
+              <TextBox>
+                <Container>
+                  <Title>기본 정보</Title>
+                  {myProfile.approval ? (
+                    <>
+                      <SUniversityMark />
+                      <UniversityMarkText>대학 인증 완료</UniversityMarkText>
+                    </>
+                  ) : (
+                    <>
+                      <SUniversityMarkGray />
+                      <UniversityNoMarkText>대학 미인증</UniversityNoMarkText>
+                    </>
+                  )}
+                </Container>
+                <TeamInfo>
+                  <Subtitle>일정</Subtitle>
+                  <SubContent>
+                    <DateText availableDates={myProfile.teamAvailableDate} />
+                  </SubContent>
+                </TeamInfo>
+                <TeamInfo>
+                  <Subtitle>지역</Subtitle>
+                  <AreaText areaProps={myProfile.areas} />
+                </TeamInfo>
+                <TeamInfo>
+                  <Subtitle>주량</Subtitle>
+                  <DrinkText drink={myProfile.drink} />
+                </TeamInfo>
+                <TeamInfo>
+                  <Subtitle>인원</Subtitle>
+                  <MemberText
+                    count={myProfile.memberCount}
+                    more={myProfile.memberCounts}
+                  />
+                </TeamInfo>
+              </TextBox>
+              <SliderBoxMembers members={myProfile.members} />
+            </TeamProfile>
           )}
         </SModal>
       ) : null}
@@ -166,21 +121,6 @@ const TeamName = styled.span`
   background-color: #ffd3d3;
   font-weight: 600;
   font-size: 16px;
-`;
-
-const EditBtn = styled.button`
-  margin-left: 15px;
-  padding: 3px 7px;
-  border: none;
-  border-radius: 5px;
-  color: #777777;
-  background-color: #e6e6e6;
-  font-size: 14px;
-`;
-
-const SInfo = styled(Info)`
-  margin-left: 10px;
-  cursor: pointer;
 `;
 
 const TextBox = styled.div`
@@ -247,26 +187,4 @@ const Subtitle = styled.span`
 const SubContent = styled.div`
   font-size: 14px;
   font-weight: 500;
-`;
-
-const Footer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 100%;
-  margin-top: 10%;
-  padding-bottom: 5%;
-`;
-
-const ButtonBox = styled.div`
-  width: 90%;
-  display: flex;
-  justify-content: center;
-  justify-content: space-between;
-  margin-top: 5%;
-`;
-
-const ApplyButton2 = styled(ApplyButton)`
-  width: 90%;
-  margin-top: 5%;
 `;
